Move dynamodb functionality to common folder
