+++ application/common/dynamodb.py
