+++ application/event_processor/tests/__init__.py
@@ -8,5 +8,4 @@ modules["dos"] = import_module("event_processor.dos")
 modules["reporting"] = import_module("event_processor.reporting")
 modules["change_event_exceptions"] = import_module("event_processor.change_event_exceptions")
 modules["change_event_validation"] = import_module("event_processor.change_event_validation")
-modules["dynamodb"] = import_module("event_processor.dynamodb")
 modules["changes"] = import_module("event_processor.changes")
